% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkerboardMass.R
\name{checkerboardDensityIndex}
\alias{checkerboardDensityIndex}
\title{Estimate a single entry of the checkerboard mass density}
\usage{
checkerboardDensityIndex(X, Y, k, l, resolution1, resolution2)
}
\arguments{
\item{X}{First coordinate of the observation.}

\item{Y}{Second coordinate of the observation.}

\item{k}{Index of the first component.}

\item{l}{Index of the second component.}

\item{resolution1}{Resolution in the first component.}

\item{resolution2}{Resolution in the second component.}
}
\value{
The estimated checkerboard mass density A_kl
}
\description{
Estimate the value \eqn{A_{kl}} of the non-square checkerboard mass density.
}
\details{
This implementation modifies the code of build_checkerboard_weights() published in 'qad', version 1.0.4, available at \href{https://CRAN.R-project.org/package=qad}{https://CRAN.R-project.org/package=qad},
to allow for the evaluation of a single index of the non-square checkerboard mass densities.
For more details on the implementation see \code{\link[qad]{ECBC}} and for more information on the implemented changes, see the file 'src/code.cpp'.
}
\examples{
checkerboardDensityIndex(runif(20), runif(20), 1, 2, 0.5, 0.5)
}
